+++ client/blueprints/tahi-task/files/__tahipluginpath__/client/app/controllers/overlays/__name__.js
@@ -1,4 +1,0 @@
-import Ember from 'ember';
-import TaskController from 'tahi/pods/paper/task/controller';
-
-export default TaskController.extend();