+++ client/blueprints/tahi-task/files/__tahipluginpath__/client/app/views/overlays/__name__.js
@@ -1,7 +1,0 @@
-import OverlayView from 'tahi/views/overlay';
-
-export default OverlayView.extend({
-  templateName: 'overlays/<%= dasherizedModuleName %>',
-  layoutName: 'layouts/overlay',
-  cardName: '<%= dasherizedModuleName %>'
-});