+++ lib/generators/tahi/task/templates/policy.rb
@@ -1,4 +1,0 @@
-module <%= @plugin_module %>
-  class <%= class_name %>TasksPolicy < ::TasksPolicy
-  end
-end